Refactor single event & subscriber
