+++ kernel/single/core/src/main/java/org/apache/shardingsphere/single/metadata/converter/SingleNodeConverter.java
@@ -37,6 +37,8 @@ public final class SingleNodeConverter {
     
     private static final String VERSION_PATTERN = "/versions/[0-9]+";
     
+    private static final String RULE_ACTIVE_VERSION = "/active_version$";
+    
     /**
      * Get tables path.
      *
@@ -59,13 +61,13 @@ public final class SingleNodeConverter {
     }
     
     /**
-     * Is tables path.
+     * Is tables active version path.
      *
      * @param rulePath rule path
      * @return true or false
      */
-    public static boolean isTablesPath(final String rulePath) {
-        Pattern pattern = Pattern.compile(RULES_NODE_PREFIX + ROOT_NODE + "/" + TABLES_NODE + VERSION_PATTERN, Pattern.CASE_INSENSITIVE);
+    public static boolean isTablesActiveVersionPath(final String rulePath) {
+        Pattern pattern = Pattern.compile(RULES_NODE_PREFIX + ROOT_NODE + "/" + TABLES_NODE + RULE_ACTIVE_VERSION, Pattern.CASE_INSENSITIVE);
         Matcher matcher = pattern.matcher(rulePath);
         return matcher.find();
     }
