+++ kernel/single/core/src/main/java/org/apache/shardingsphere/single/subscriber/SingleConfigurationSubscriber.java
@@ -22,12 +22,14 @@ import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.instance.InstanceContext;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
 import org.apache.shardingsphere.infra.rule.RuleConfigurationSubscribeCoordinator;
+import org.apache.shardingsphere.infra.util.yaml.YamlEngine;
 import org.apache.shardingsphere.mode.event.config.DatabaseRuleConfigurationChangedEvent;
 import org.apache.shardingsphere.single.api.config.SingleRuleConfiguration;
 import org.apache.shardingsphere.single.event.config.AddSingleTableEvent;
 import org.apache.shardingsphere.single.event.config.AlterSingleTableEvent;
 import org.apache.shardingsphere.single.event.config.DeleteSingleTableEvent;
 import org.apache.shardingsphere.single.rule.SingleRule;
+import org.apache.shardingsphere.single.yaml.config.pojo.YamlSingleRuleConfiguration;
 
 import java.util.Map;
 import java.util.Optional;
@@ -58,7 +60,8 @@ public final class SingleConfigurationSubscriber implements RuleConfigurationSub
     @Subscribe
     public synchronized void renew(final AddSingleTableEvent event) {
         ShardingSphereDatabase database = databases.get(event.getDatabaseName());
-        SingleRuleConfiguration needToAddedConfig = event.getConfig();
+        SingleRuleConfiguration needToAddedConfig = swapSingleTableRuleConfig(
+                instanceContext.getModeContextManager().getVersionPathByActiveVersionKey(event.getActiveVersionKey(), event.getActiveVersion()));
         Optional<SingleRule> rule = database.getRuleMetaData().findSingleRule(SingleRule.class);
         SingleRuleConfiguration config;
         if (rule.isPresent()) {
@@ -79,7 +82,8 @@ public final class SingleConfigurationSubscriber implements RuleConfigurationSub
     @Subscribe
     public synchronized void renew(final AlterSingleTableEvent event) {
         ShardingSphereDatabase database = databases.get(event.getDatabaseName());
-        SingleRuleConfiguration needToAlteredConfig = event.getConfig();
+        SingleRuleConfiguration needToAlteredConfig = swapSingleTableRuleConfig(
+                instanceContext.getModeContextManager().getVersionPathByActiveVersionKey(event.getActiveVersionKey(), event.getActiveVersion()));
         SingleRuleConfiguration config = database.getRuleMetaData().getSingleRule(SingleRule.class).getConfiguration();
         config.setTables(needToAlteredConfig.getTables());
         config.setDefaultDataSource(needToAlteredConfig.getDefaultDataSource().orElse(null));
@@ -98,5 +102,15 @@ public final class SingleConfigurationSubscriber implements RuleConfigurationSub
         config.getTables().clear();
         config.setDefaultDataSource(null);
         instanceContext.getEventBusContext().post(new DatabaseRuleConfigurationChangedEvent(event.getDatabaseName(), config));
+    }
+    
+    private SingleRuleConfiguration swapSingleTableRuleConfig(final String yamlContext) {
+        SingleRuleConfiguration result = new SingleRuleConfiguration();
+        YamlSingleRuleConfiguration yamlSingleRuleConfiguration = YamlEngine.unmarshal(yamlContext, YamlSingleRuleConfiguration.class);
+        if (null != yamlSingleRuleConfiguration.getTables()) {
+            result.getTables().addAll(yamlSingleRuleConfiguration.getTables());
+        }
+        result.setDefaultDataSource(yamlSingleRuleConfiguration.getDefaultDataSource());
+        return result;
     }
 }
